requirements.txt
redis>=4.x.x
pyyaml>=6.0
